agent_financial_advice/requirements.txt
# Anthropic Claude API
anthropic>=0.30.0

# Market data
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0

# News & web fetching
feedparser>=6.0.11
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Config & environment
python-dotenv>=1.0.1
PyYAML>=6.0.1

# Newsletter formatting
markdown>=3.6

# Scheduling
apscheduler>=3.10.4

# Email delivery
sendgrid>=6.11.0

# WhatsApp delivery
twilio>=9.2.0

# Console output
rich>=13.7.0
